requirements.txt
google-cloud-documentai>=2.20.0
google-cloud-core>=2.4.1
google-api-core>=2.17.1
pandas>=2.0.0
pathlib2>=2.3.7; python_version < '3.4' 